Float is DOUBLE now

+++ codigo/include/criterios.h
@@ -4,25 +4,25 @@
 class Criterios{
 	public:
 		Criterios(){}
-		Criterios(int x, float e):max_iter(x),epsilon(e){}
+		Criterios(int x, double e):max_iter(x),epsilon(e){}
 
 		//N = max_iter 
 		bool criterio1_parar(int x){return x >= max_iter;}
 
 		// |x - x_anterior| < epsilon
-		bool criterio2_parar(float x, float x_anterior);
+		bool criterio2_parar(double x, double x_anterior);
 
 		// ( |x - x_anterior| / |x_anterior| ) < epsilon
-		bool criterio2_bis_parar(float x, float x_anterior);
+		bool criterio2_bis_parar(double x, double x_anterior);
 
 		// |f(x)| < epsilon
-		bool criterio3_parar(float y);
+		bool criterio3_parar(double y);
 
 		// |f(x) - f(x_anterior)| < epsilon
-		bool criterio4_parar(float y, float y_anterior);
+		bool criterio4_parar(double y, double y_anterior);
 
 		// ( |f(x) - f(x_anterior)| / |f(x_anterior)| ) < epsilon
-		bool criterio4_bis_parar(float y, float y_anterior);
+		bool criterio4_bis_parar(double y, double y_anterior);
 		
 		//	Todos los criterios en 1, con selector se elige cual:
 		//	1:Maximo Iteraciones
@@ -31,11 +31,11 @@ class Criterios{
 		//	4:Error absoluto
 		//	5:Distancia absoluta de valores de iteraciones sucesivas
 		//	6:Distancia relativa de valores de iteraciones sucesivas
-		bool criterios(int selector, int maxIter,float x,float x_anterior,float f_x,float f_x_anterior);
+		bool criterios(int selector, int maxIter,double x,double x_anterior,double f_x,double f_x_anterior);
 
 	private:
 		int max_iter;
-		float epsilon;
+		double epsilon;
 };
 
 #endif
