+++ codigo/include/funciones.h
@@ -9,24 +9,24 @@ class Funciones{
 
 	public:
 		Funciones(){}
-		Funciones(float x){alpha = x;} 
+		Funciones(double x){alpha = x;} 
 
 		// f corresponde al primer experimento para obtener sqrt(alpha) 
-		float f(float x){return x*x - this->alpha;}
-		float f_deriv(float x){return 2.0*x;}
-		pair<float,float> semillas_biseccion_f(int);
+		double f(double x){return x*x - this->alpha;}
+		double f_deriv(double x){return 2.0*x;}
+		pair<double,double> semillas_biseccion_f(int);
 
 		// e corresponde al segundo experimento para obtener 1/sqrt(alpha)
-		float e(float x){return 1.0/(x*x) - this->alpha;}		
-		float e_deriv(float x){ return (-2.0)/(x*x*x);}
-		pair<float,float> semillas_biseccion_e(int);
+		double e(double x){return 1.0/(x*x) - this->alpha;}		
+		double e_deriv(double x){ return (-2.0)/(x*x*x);}
+		pair<double,double> semillas_biseccion_e(int);
 
-		float cero_teorico(){return sqrt(alpha);}
-		float cero_teorico_e(){return 1.0/sqrt(alpha);}
-		float get_alpha(){return alpha;}
+		double cero_teorico(){return sqrt(alpha);}
+		double cero_teorico_e(){return 1.0/sqrt(alpha);}
+		double get_alpha(){return alpha;}
 
 	private:
-		float alpha;
+		double alpha;
 };
 
 #endif
